Rest Location complete

+++ src/main/java/com/merinosa/geolocation/vjsantojaca/server/models/entities/LocationEntity.java
@@ -27,10 +27,10 @@ public class LocationEntity implements Serializable {
 	private int idUser;
 	
 	@Column(name="latitude")
-	private float latitude;
+	private double latitude;
 	
 	@Column(name="longitude")
-	private float longitude;
+	private double longitude;
 	
 	@Column(name="date")
 	private long date;	
@@ -66,19 +66,19 @@ public class LocationEntity implements Serializable {
 		this.idUser = idUser;
 	}
 
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
